--- application/src/scouter/input-types/auto-map/AutonomousForm.tsx
@@ -1,44 +1,0 @@
-import React, { useState } from "react";
-import SushiButton from "./SushiButton";
-import {
-  AllSushis,
-  AllSushisAndStorage,
-  Sushi,
-  ValuesToBePassed,
-} from "./AutonomousMapInput";
-import { SushiToBeChanged } from "./AutonomousMapInput";
-
-const AutonomousForm: React.FC<AllSushisAndStorage> = (props) => {
-  const sushiButtons = [
-    <SushiButton
-      sushies={props.allSushis}
-      sushiToBeChanged={"Sushi1"}
-      storage={props.storage}
-    />,
-    <SushiButton
-      sushies={props.allSushis}
-      sushiToBeChanged={"Sushi2"}
-      storage={props.storage}
-    />,
-    <SushiButton
-      sushies={props.allSushis}
-      sushiToBeChanged={"Sushi3"}
-      storage={props.storage}
-    />,
-  ];
-  const buttons = <div className="sushi-buttons">{sushiButtons}</div>;
-
-  const blueAllienceAutonomousMap = (
-    <div className="field-container">
-      <img
-        src="/src/assets/blue-auto-map.png"
-        alt="Field"
-        style={{ width: "70%" }}
-      ></img>
-      {buttons}
-    </div>
-  );
-
-  return <>{blueAllienceAutonomousMap}</>;
-};
-export default AutonomousForm;
